LaV2Gen/requirements.txt
numpy==1.21.2
torch==1.9.1
pandas==1.3.2
# jury==2.1.0
tqdm==4.62.3
torchvision==0.10.1
# Shapely==1.7.1
transformers==4.11.3
Pillow==8.4.0
scikit_learn==1.0.1
tensorboardX==2.4
